feat(mha): half of mha refactor code
