+++ internlm/model/modeling_internlm2.py
@@ -3,8 +3,6 @@ import math
 from typing import Optional
 
 import torch
-import torch.nn.functional as F
-from einops import rearrange
 from torch import nn
 
 from internlm.core.context import ParallelMode
@@ -15,19 +13,11 @@ from internlm.initialize.initialize_tensor import (
     scaled_init_method_uniform,
     uniform_,
 )
-from internlm.model.embedding import (
-    DynamicNTKScalingRotaryEmbedding,
-    Embedding1D,
-    RotaryEmbedding,
-)
+from internlm.model.embedding import Embedding1D
+from internlm.model.modules.multi_head_attention import MHA3
 from internlm.model.modules.ffn import new_fead_forward
 from internlm.model.modules.linear import new_linear
-from internlm.model.multi_head_attention import (
-    CrossAttention,
-    DistributedAttention,
-    SelfAttention,
-    _update_kv_cache,
-)
+
 from internlm.model.modules.utils import (
     split_forward_gather_backward,
     try_import_RMSNorm,
@@ -42,415 +32,6 @@ MODEL_TYPE = "INTERNLM2_PUBLIC"
 
 logger = get_logger(__file__)
 RMSNorm = try_import_RMSNorm()
-
-
-class MHA(nn.Module):
-    """
-    Multi-head self-attention and cross-attention.
-
-    Args:
-        embed_dim (int): The dimention of hidden state.
-        num_heads (int): The number of attention heads.
-        num_kv_heads (int): The number of attention heads for key and value.
-        process_group (torch.distributed.ProcessGroup): The group of the current device for `parallel_mode`.
-        sequence_process_group (torch.distributed.ProcessGroup): The process group for attention calculation.
-        bias (bool): Whether the bias is needed for linears. Will be used when initializing QKV matrix and
-                     output projection. False by default.
-        dropout (float): The dropout rate for cross attention and self attention. 0.0 by default.
-        softmax_scale (float): The temperature to use for the softmax attention.
-        causal (boolean): Whether to apply causal attention mask. False by default.
-        layer_idx (int): The index of current layer. None by default.
-        rope_base (int): The value of `base` for rotary position embeddings. 10000 by default.
-        rotary_emb_dim (int): The dimention of Rotary Embedding. 0 by default.
-        rotary_emb_scale_base (int): The scaling factor of Rotary Embedding. If scale_base > 0, this implements
-                                    XPos(Sun et al., https://arxiv.org/abs/2212.10554). 0 by default.
-        use_flash_attn (bool): Whether to use flash attention or not.If False, vanilla attention module will be used.
-                               False by default.
-        device (Optional[Union[str, torch.device]]): The device will be used.
-        dtype (Optional[torch.dtype]): The type of data.
-        rot_embed_HF_impl (Optional[bool]): Whether to use the rotary embedding implementation from HuggingFace.
-                                            True by default.
-        tp_mode (str): The string value of tensor parallel mode, should be in ["mtp", "msp", "fsp", "isp"],
-                       "mtp" by default.
-    """
-
-    def __init__(
-        self,
-        embed_dim: int,
-        num_heads: int,
-        num_kv_heads: int,
-        process_group: Optional[torch.distributed.ProcessGroup],
-        sequence_process_group: Optional[torch.distributed.ProcessGroup],
-        max_position_embeddings: int = 2048,
-        bias: bool = False,
-        dropout: float = 0.0,
-        softmax_scale: float = None,
-        causal: bool = False,
-        layer_idx: int = None,
-        use_dynamic_ntk_rope: bool = False,
-        use_flash_attn: bool = True,
-        rope_base: int = 10000,
-        rotary_emb_dim: int = 0,
-        rotary_emb_scale_base: int = 0,
-        device: Optional[torch.device] = None,
-        dtype: Optional[torch.dtype] = None,
-        rot_embed_HF_impl: Optional[bool] = True,
-        tp_mode: str = "mtp",
-    ) -> None:
-        factory_kwargs = {"device": device, "dtype": dtype}
-        super().__init__()
-        self.embed_dim = embed_dim
-        self.num_heads = num_heads
-        assert self.embed_dim % num_heads == 0, "embedding dim must be divisible by num_heads"
-
-        self.head_dim = self.embed_dim // num_heads
-        self.num_kv_heads = num_kv_heads
-        self.kv_dim = self.head_dim * num_kv_heads
-        self.causal = causal
-        self.layer_idx = layer_idx
-        self.rotary_emb_dim = rotary_emb_dim
-        self.use_flash_attn = use_flash_attn
-        self.dtype = dtype
-
-        self.q_per_kv = num_heads // num_kv_heads
-
-        self.rot_embed_HF_impl = rot_embed_HF_impl
-
-        self.max_position_embeddings = max_position_embeddings
-        self.use_dynamic_ntk_rope = use_dynamic_ntk_rope
-        self.tp_mode = tp_mode
-
-        if self.rotary_emb_dim > 0:
-            if self.use_dynamic_ntk_rope:
-                self.rotary_emb = DynamicNTKScalingRotaryEmbedding(
-                    self.rotary_emb_dim,
-                    base=rope_base,
-                    scale_base=rotary_emb_scale_base,
-                    device=device,
-                    max_position_embeddings=max_position_embeddings,
-                    scaling_factor=1.0,  # Currently do not support dynamic scaling.
-                )
-            else:
-                self.rotary_emb = RotaryEmbedding(
-                    self.rotary_emb_dim, base=rope_base, scale_base=rotary_emb_scale_base, device=device
-                )
-
-        self.wqkv = new_linear("wqkv", embed_dim, embed_dim + 2 * self.kv_dim, bias, **factory_kwargs)
-
-        if use_flash_attn:
-            from flash_attn import flash_attn_varlen_kvpacked_func
-            from flash_attn.modules.mha import FlashCrossAttention, FlashSelfAttention
-
-        inner_attn_cls = FlashSelfAttention if use_flash_attn else SelfAttention
-        inner_cross_attn_cls = FlashCrossAttention if use_flash_attn else CrossAttention
-        self.inner_attn = inner_attn_cls(causal=causal, softmax_scale=softmax_scale, attention_dropout=dropout)
-        self.inner_cross_attn = inner_cross_attn_cls(
-            causal=causal, softmax_scale=softmax_scale, attention_dropout=dropout
-        )
-
-        self.inner_cross_attn_causal = causal
-        self.inner_cross_attn_softmax_scale = softmax_scale
-        self.inner_cross_attn_dropout = dropout
-
-        self.attn = flash_attn_varlen_kvpacked_func if use_flash_attn else SelfAttention
-        if self.tp_mode == "isp":
-            self.attn = DistributedAttention(self.attn, sequence_process_group=sequence_process_group)
-
-        self.wo = new_linear("wo", embed_dim, embed_dim, bias, **factory_kwargs)
-
-    def forward(self, x, seqlen=None, inference_params=None, **kwargs):
-        if kwargs.get("indexes", None) is not None:
-            return self._packed_forward(x=x, inference_params=inference_params, **kwargs)
-        else:
-            return self._forward(x=x, seqlen=seqlen, inference_params=inference_params, **kwargs)
-
-    def _forward(self, x, seqlen=None, inference_params=None, **kwargs):  # pylint: disable=W0613
-        """
-        Arguments:
-            x: (batch, seqlen, hidden_dim) (where hidden_dim = num heads * head dim) if seqlen=None.
-                If seqlen is not None, x is (batch * seqlen, hidden_dim). This is so that when we
-                split x during sequence parallel, we split the batch * seqlen dimension
-                (in case batch is small).
-        """
-        bsz, _, _ = x.shape
-        qkv = self.wqkv(x)
-
-        if seqlen is None:
-            qkv = rearrange(qkv, "b s (h gs d) -> b s h gs d", gs=self.q_per_kv + 2, d=self.head_dim)
-        else:
-            qkv = rearrange(qkv, "(b s) (h gs d) -> b s h gs d", s=seqlen, gs=self.q_per_kv + 2, d=self.head_dim)
-
-        q, k, v = (qkv[..., : self.q_per_kv, :], qkv[..., -2, :], qkv[..., -1, :])
-
-        q = rearrange(q, "b s h gs d -> b s (h gs) d")
-
-        if not self.rot_embed_HF_impl:
-            q = torch.cat([q[..., ::2], q[..., 1::2]], dim=-1)
-            k = torch.cat([k[..., ::2], k[..., 1::2]], dim=-1)
-
-        if inference_params is None:
-            if self.rotary_emb_dim > 0:
-                q = self.rotary_emb._single_eval_forward(q)
-                k = self.rotary_emb._single_eval_forward(k)
-            kv = torch.concat([k.unsqueeze(2), v.unsqueeze(2)], dim=2)
-            if self.dtype is torch.float32 and self.use_flash_attn:
-                if q.dtype not in [torch.float16, torch.bfloat16]:
-                    q = q.to(torch.bfloat16)
-                if kv.dtype not in [torch.float16, torch.bfloat16]:
-                    kv = kv.to(torch.bfloat16)
-                with torch.cuda.amp.autocast(dtype=torch.bfloat16):
-                    context = self.inner_cross_attn(q, kv).to(self.dtype)
-            else:
-                context = self.inner_cross_attn(q, kv)
-
-        else:
-            assert self.rotary_emb_dim > 0
-            if hasattr(inference_params, "attention_mask") and inference_params.attention_mask is not None:
-                empties = inference_params.attention_mask[..., -1].sum(dim=-1)
-                moved_q = q.clone()
-                moved_k = k.clone()
-                if inference_params.sequence_len_offset == 0:
-                    for i in range(len(empties)):
-                        if empties[i] != 0:
-                            moved_q[i][: -empties[i]] = q[i][empties[i] :]
-                            moved_k[i][: -empties[i]] = k[i][empties[i] :]
-                    moved_q = self.rotary_emb._single_eval_forward(
-                        moved_q, seqlen_offset=inference_params.sequence_len_offset
-                    )
-                    moved_k = self.rotary_emb._single_eval_forward(
-                        moved_k, seqlen_offset=inference_params.sequence_len_offset
-                    )
-                    for i in range(len(empties)):
-                        if empties[i] != 0:
-                            q[i][empties[i] :] = moved_q[i][: -empties[i]]
-                            k[i][empties[i] :] = moved_k[i][: -empties[i]]
-                        else:
-                            q[i] = moved_q[i]
-                            k[i] = moved_k[i]
-                else:
-                    q = q.squeeze(1)
-                    k = k.squeeze(1)
-                    cu_seqlens = kwargs.get("cu_seqlens", None)
-                    max_seqlen = kwargs.get("max_seqlen", None)
-                    q = self.rotary_emb._single_forward(
-                        q,
-                        inference_params.sequence_len_offset * torch.ones(q.size(0), dtype=torch.int, device=q.device)
-                        - empties,
-                        cu_seqlens=cu_seqlens,
-                        max_seqlen=max_seqlen,
-                    ).unsqueeze(1)
-                    k = self.rotary_emb._single_forward(
-                        k,
-                        inference_params.sequence_len_offset * torch.ones(k.size(0), dtype=torch.int, device=k.device)
-                        - empties,
-                        cu_seqlens=cu_seqlens,
-                        max_seqlen=max_seqlen,
-                    ).unsqueeze(1)
-            else:
-                raise NotImplementedError(
-                    "You should make sure you are aware that you are changing the method of generating."
-                    "According to your generation function instead of inference/seq_generator_module.py, "
-                    "You may implement here for normal running."
-                )
-
-            kv = torch.stack([k, v], dim=2)
-
-            assert self.layer_idx is not None, "Generation requires layer_idx in the constructor"
-            if hasattr(inference_params, "window_size") and inference_params.window_size is not None:
-                if inference_params.window_size <= inference_params.sequence_len_offset:
-                    assert kv.size(1) == 1, "update kv lenth more than 1"
-                    inference_params.key_value_memory_dict[self.layer_idx][
-                        :, inference_params.keep_first : inference_params.window_size - 1, ...
-                    ] = inference_params.key_value_memory_dict[self.layer_idx][
-                        :, -(inference_params.window_size - 1 - inference_params.keep_first) :, ...
-                    ].clone()
-                    inference_params.real_sequence_len_offset = inference_params.sequence_len_offset
-                    inference_params.sequence_len_offset = inference_params.window_size - 1
-
-                    kv = _update_kv_cache(kv, inference_params, self.layer_idx)
-
-                    inference_params.sequence_len_offset = inference_params.real_sequence_len_offset
-                else:
-                    kv = _update_kv_cache(kv, inference_params, self.layer_idx)
-            else:
-                kv = _update_kv_cache(kv, inference_params, self.layer_idx)
-
-            # When using FP16, there is a high probability of NAN in the KV.
-            # Since NAN cannot be removed by multiplying with and 0, it needs
-            # to be removed manually here.
-            kv = torch.where(torch.isnan(kv), 0, kv)
-
-            if hasattr(inference_params, "attention_mask") and inference_params.attention_mask is not None:
-                assert self.use_flash_attn is True
-                from flash_attn import flash_attn_varlen_kvpacked_func
-
-                if inference_params.sequence_len_offset == 0:  # First entrance, attnmask (bs*seqlen*seqlen)
-                    attn_mask = inference_params.attention_mask[:, None, ...]
-                    attn_mask = torch.logical_or(
-                        torch.ones_like(attn_mask, dtype=torch.bool).triu(diagonal=1), attn_mask
-                    )
-                    attn_mask4flsh = ~attn_mask[:, :, -1, :].view(bsz, -1)
-                    cu_seqlens = torch.concat(
-                        [
-                            torch.tensor([0], dtype=torch.int32, device=attn_mask4flsh.device),
-                            attn_mask4flsh.sum(dim=-1).to(dtype=torch.int32),
-                        ],
-                        dim=0,
-                    )
-                    cu_seqlens = cu_seqlens.cumsum(dim=0, dtype=torch.int32)
-                    max_seqlen_q = attn_mask4flsh.shape[-1]
-                    max_seqlen_k = attn_mask4flsh.shape[-1]
-                    total_q = q.masked_select(attn_mask4flsh.view(bsz, -1, 1, 1)).view(-1, q.shape[-2], q.shape[-1])
-                    total_kv = kv.masked_select(attn_mask4flsh.view(bsz, -1, 1, 1, 1)).view(
-                        -1, kv.shape[-3], kv.shape[-2], kv.shape[-1]
-                    )
-                    if self.dtype is torch.float32:
-                        if total_q.dtype not in [torch.float16, torch.bfloat16]:
-                            total_q = total_q.to(torch.bfloat16)
-                        if total_kv.dtype not in [torch.float16, torch.bfloat16]:
-                            total_kv = total_kv.to(torch.bfloat16)
-                        with torch.cuda.amp.autocast(dtype=torch.bfloat16):
-                            output = flash_attn_varlen_kvpacked_func(
-                                q=total_q,
-                                kv=total_kv,
-                                cu_seqlens_q=cu_seqlens,
-                                cu_seqlens_k=cu_seqlens,
-                                max_seqlen_q=max_seqlen_q,
-                                max_seqlen_k=max_seqlen_k,
-                                dropout_p=0.0,
-                                causal=True,
-                            ).to(self.dtype)
-                    else:
-                        output = flash_attn_varlen_kvpacked_func(
-                            q=total_q,
-                            kv=total_kv,
-                            cu_seqlens_q=cu_seqlens,
-                            cu_seqlens_k=cu_seqlens,
-                            max_seqlen_q=max_seqlen_q,
-                            max_seqlen_k=max_seqlen_k,
-                            dropout_p=0.0,
-                            causal=True,
-                        )
-
-                    context = torch.zeros_like(q)
-                    context = context.masked_scatter_(attn_mask4flsh.view(bsz, -1, 1, 1), output)
-
-                else:
-                    attn_mask = inference_params.attention_mask[:, -1, :].view(bsz, 1, 1, -1)
-                    if hasattr(inference_params, "window_size") and inference_params.window_size is not None:
-                        if inference_params.window_size <= inference_params.sequence_len_offset:
-                            attn_mask = torch.concat(
-                                [
-                                    attn_mask[..., : inference_params.keep_first],
-                                    attn_mask[..., -(inference_params.window_size - inference_params.keep_first) :],
-                                ],
-                                dim=-1,
-                            )
-
-                    k, v = torch.chunk(kv, 2, dim=2)
-                    k = k.squeeze(2)
-                    v = v.squeeze(2)
-                    sp = k.shape
-                    expansion = q.size(2) // k.size(2)
-                    scores = torch.einsum(
-                        "blhd,bnhd->bhln",
-                        q,
-                        k.unsqueeze(3).expand(-1, -1, -1, expansion, -1).reshape(sp[0], sp[1], q.size(2), sp[3]),
-                    ) / math.sqrt(q.size(-1))
-                    scores = scores.masked_fill(attn_mask, -65000.0)
-                    scores = F.softmax(scores, dim=-1)  # bsz x h x L x L
-                    context = torch.einsum(
-                        "bhmn,bnhd->bmhd",
-                        scores,
-                        v.unsqueeze(3).expand(-1, -1, -1, expansion, -1).reshape(sp[0], sp[1], q.size(2), sp[3]),
-                    )
-            else:
-                if self.dtype is torch.float32 and self.use_flash_attn:
-                    if q.dtype not in [torch.float16, torch.bfloat16]:
-                        q = q.to(torch.bfloat16)
-                    if kv.dtype not in [torch.float16, torch.bfloat16]:
-                        kv = kv.to(torch.bfloat16)
-                    with torch.cuda.amp.autocast(dtype=torch.bfloat16):
-                        context = self.inner_cross_attn(q, kv, causal=True).to(self.dtype)
-                else:
-                    context = self.inner_cross_attn(q, kv, causal=True)
-
-        if seqlen is None:
-            context = rearrange(context, "b s h d -> b s (h d)")
-        else:
-            context = rearrange(context, "b s h d -> (b s) (h d)")
-
-        out = self.wo(context)
-        return out
-
-    def _packed_forward(self, x, inference_params=None, **kwargs):
-        """
-        Arguments:
-            x: (batch, seqlen, hidden_dim) (where hidden_dim = num heads * head dim) if seqlen=None.
-                If seqlen is not None, x is (batch * seqlen, hidden_dim). This is so that when we
-                split x during sequence parallel, we split the batch * seqlen dimension
-                (in case batch is small).
-        """
-        assert self.use_flash_attn is True
-
-        qkv = self.wqkv(x)
-
-        qkv = rearrange(qkv, "t (h gs d) -> t h gs d", gs=self.q_per_kv + 2, d=self.head_dim)
-
-        q, k, v = (qkv[..., : self.q_per_kv, :], qkv[..., -2, :], qkv[..., -1, :])
-
-        q = rearrange(q, "t h gs d -> t (h gs) d")
-
-        # qkv shift
-        # the rotary embedding in flash attention module in performed by separating the front and back parts, while
-        # most of others are done by odd-even methods.
-        if not self.rot_embed_HF_impl:
-            q = torch.cat([q[..., ::2], q[..., 1::2]], dim=-1)
-            k = torch.cat([k[..., ::2], k[..., 1::2]], dim=-1)
-
-        indexes = kwargs.pop("indexes")
-        cu_seqlens = kwargs.pop("cu_seqlens")
-        max_seqlen = kwargs.pop("max_seqlen")
-        q = self.rotary_emb._single_forward(q, indexes=indexes, cu_seqlens=cu_seqlens, max_seqlen=max_seqlen)
-        k = self.rotary_emb._single_forward(k, indexes=indexes, cu_seqlens=cu_seqlens, max_seqlen=max_seqlen)
-
-        if inference_params is None:
-            kv = torch.concat([k.unsqueeze(1), v.unsqueeze(1)], dim=1)
-            if self.dtype is torch.float32:
-                if q.dtype not in [torch.float16, torch.bfloat16]:
-                    q = q.to(torch.bfloat16)
-                if kv.dtype not in [torch.float16, torch.bfloat16]:
-                    kv = kv.to(torch.bfloat16)
-                with torch.cuda.amp.autocast(dtype=torch.bfloat16):
-                    context = self.attn(
-                        q=q,
-                        kv=kv,
-                        cu_seqlens_q=kwargs["cu_seqlens"],
-                        cu_seqlens_k=kwargs["cu_seqlens"],
-                        max_seqlen_q=kwargs["max_seqlen"],
-                        max_seqlen_k=kwargs["max_seqlen"],
-                        dropout_p=self.inner_cross_attn_dropout,
-                        softmax_scale=self.inner_cross_attn_softmax_scale,
-                        causal=self.inner_cross_attn_causal,
-                    ).to(self.dtype)
-            else:
-                context = self.attn(
-                    q=q,
-                    kv=kv,
-                    cu_seqlens_q=kwargs["cu_seqlens"],
-                    cu_seqlens_k=kwargs["cu_seqlens"],
-                    max_seqlen_q=kwargs["max_seqlen"],
-                    max_seqlen_k=kwargs["max_seqlen"],
-                    dropout_p=self.inner_cross_attn_dropout,
-                    softmax_scale=self.inner_cross_attn_softmax_scale,
-                    causal=self.inner_cross_attn_causal,
-                )
-        else:
-            raise RuntimeError("Not support this right now")
-
-        context = rearrange(context, "b h d -> b (h d)")  # recover shape
-        out = self.wo(context)
-        return out
 
 
 class PackedFlashLlamaLayer1D(nn.Module):
@@ -537,7 +118,7 @@ class PackedFlashLlamaLayer1D(nn.Module):
         parallel_mode = ParallelMode.WEIGHT if self.tp_mode == "isp" else ParallelMode.TENSOR
 
         head_dim = hidden_size // num_attention_heads
-        self.attention = MHA(
+        self.attention = MHA3(
             embed_dim=hidden_size,
             num_heads=num_attention_heads,
             num_kv_heads=num_kv_attention_heads,
