+++ internlm/model/modeling_moe.py
@@ -15,7 +15,7 @@ from internlm.model.embedding import Embedding1D
 from internlm.model.modules.ffn import new_fead_forward
 from internlm.model.modules.linear import new_linear
 from internlm.model.moe import MoE
-from internlm.model.multi_head_attention import MHA
+from internlm.model.modules.multi_head_attention import MHA
 from internlm.model.modules.utils import (
     split_forward_gather_backward,
     try_import_RMSNorm,
